LoongArch: Add efistub booting support
This patch adds efistub booting support, which is the standard UEFI boot
protocol for us to use.

We use generic efistub, which means we can pass boot information (i.e.,
system table, memory map, kernel command line, initrd) via a light FDT
and drop a lot of non-standard code.

We use a flat mapping to map the efi runtime in the kernel's address
space. In efi, VA = PA; in kernel, VA = PA + PAGE_OFFSET. As a result,
flat mapping is not identity mapping, SetVirtualAddressMap() is still
needed for the efi runtime.

Currently, generic efistub doesn't support mapping efi runtime in the
kernel. So we set efi_novamap to not call SetVirtualAddressMap() in the
stub. Instead, we call it in the core kernel. This also makes the raw
elf kernel booting be possible, which is needed by non-UEFI firmware
(e.g., PMON which is widely used by Loongson for historic reasons).

Then how the elf kernel and the efi kernel co-exist? When building, the
raw vmlinux is naturally in elf format, the efi kernel is generated from
vmlinux by objcopy via removing the elf header.

Note: The magic number in MSDOS header is used by Grub [1], which is the
same as RISC-V and ARM64.

[1] https://lists.gnu.org/archive/html/grub-devel/2021-10/msg00215.html

Signed-off-by: Huacai Chen <chenhuacai@loongson.cn>
(cherry picked from commit 5d5ea41ec2ce90385165efd38f75cd6ec4748acb)