LoongArch: Add section of GOT for kernel module
The address of external symbols will locate more than 32-bit offset.  We
were using the `-Wa,-mla-global-with-abs` and `-Wa,-mla-local-with-abs`
to prevent the compiler and assembler from generating GOT relocations,
but these options are undocumented hacks and do not work anymore with
GAS 2.40 and GCC 13.

Let the module loader emit GOT entries for data symbols so we would be
able to handle GOT relocations.  The GOT entry is just the data symbol
address.

In module.lds, emit a stub .got section for a section header entry.
The actual content of the entry will be filled at runtime by
module_frob_arch_sections.

A special symbol named "_GLOBAL_OFFSET_TABLE_" is used by stack-based
relocations for the PC-relative offset of a GOT entry, like:

    R_LARCH_SOP_PUSH_PCREL _GLOBAL_OFFSET_TABLE_
    R_LARCH_SOP_PUSH_GPREL foo
    R_LARCH_SOP_ADD

Each kernel module has its own GOT (like a shared object), so we need
to generate _GLOBAL_OFFSET_TABLE_ as a local symbol for each module.

Signed-off-by: Xi Ruoyao <xry111@xry111.site>