LoongArch: Remove -fplt and -Wa,-mla-* from CFLAGS
-fplt is the default of all supported compilers (GCC, and maybe Clang in
the future), so it needs not to be explicitly specified.

-Wa,-mla-* options were used to prevent the assembler from generating
GOT accesses for "la.local" and "la.global" macros.  But GCC >= 13 will
generate GOT access explicitly without use of these macros by default if
the assembler supports explicit specification for relocations.  And,
it's not recommended to override the default as the use of assembler
macros may limit optimization.

Now we can handle GOT and GOT-based relocations properly, so we can
remove these options and use GOT access for both GCC 12 and 13
(or newer).

GAS <= 2.39 does not support "la.got [reg], [sym] + [offset]" with a
non-zero offset.  So in the assembly code we explicitly use "la.pcrel"
instead of "la" (now defaulted to "la.got") where a PC-relative
addressing is suitable, in order to work around this limitation and keep
the compatibility with old toolchains.

Signed-off-by: Xi Ruoyao <xry111@xry111.site>