LoongArch: use static-PIE-style LDFLAGS for CONFIG_RELOCATABLE instead of shared-object-style
A userspace static PIE is linked with

    -static -pie --no-dynamic-linker -z text

The kernel is more similar to static PIE, instead of shared object.  But
we need to use -z notext here.  I guess the reason is the kernel doesn't
really enforce RWX permissions for kernel pages and some relocations in
read-only segments are used deliberately.

Signed-off-by: Xi Ruoyao <xry111@xry111.site>