arm64: Add overrride for MPAM
As the message of the commit 09e6b306f3ba ("arm64: cpufeature: discover
CPU support for MPAM"), if a buggy firmware fails to either enable MPAM
or emulate the trap as if it were disabled, the kernel will just fail to
boot.  While upgrading the firmware should be the best solution, we have
some hardware of which the vendor have no response after one month we
requested a firmware update.  Allow overriding it so our devices don't
become e-waste.

Signed-off-by: Xi Ruoyao <xry111@xry111.site>